Bring m7-08-sweep (M7+M7b+M7c 0.8.x burn-down, ~70 items, 0.7.0 ship
state) and the 0.8.0 desktop wave (DESIGN.md v0.1 + canonical
sidebar + 12 app stubs + audit ratchet + 14 Plane items) onto master
in a single merge.

This is a local merge to evaluate how the 323-commit divergence
lands on master. Not tagged, not pushed. Reviewable as the cumulative
state of two completed waves (0.7.0 release-gate + 0.8.0 desktop
foundation).

Headline contents:

m7-08-sweep:
- 0.7.0 release-gate audit closure: 74 items shipped, 14 new items
  filed, 3 security highs patched (WEFT-569 URL token via fragment,
  WEFT-570 token revoke RPC, WEFT-576 nginx-unprivileged image).
- Full build green: native, browser-WASM (panel + clawft-wasm), WASI,
  React UI. Workspace check + clippy + 1494 tests pass.
- 18 audit clusters verified, 14 follow-up Plane items filed for 0.9.x.

0.8.0 desktop wave (this branch's contribution):
- docs/DESIGN.md (v0.1, ~470 lines) — design contract.
- .claude/skills/weftos-design/ — operational skill + 3 audit scripts.
- 13 hi-fi mockups (base view + 12 app states).
- crates/clawft-gui-egui/src/shell/sidebar.rs — canonical sidebar.
- crates/clawft-gui-egui/src/apps/ — 12 app stubs + state helper +
  dispatch.
- crates/clawft-gui-egui/src/theming.rs — bg_sidebar token + token
  contract test (catches DESIGN.md ↔ runtime drift).
- audit-theme.sh --baseline ratchet + pr-gates CI gate.
- 14 Plane items filed: WEFT-578..591.

Validation at the merge tip:
- scripts/build.sh check + clippy clean.
- cargo test -p clawft-gui-egui --lib: 337 / 337 pass.
- audit-theme.sh --baseline: holds at 246 offenders.

+++ crates/clawft-graphify/src/report.rs
@@ -350,7 +350,7 @@ mod tests {
             target: EntityId::new(&DomainTag::Code, &EntityType::Function, tgt_name, tgt_file),
             relation_type: RelationType::Calls,
             confidence: conf,
-            weight: conf.to_weight() as f32,
+            weight: conf.to_weight(),
             source_file: None,
             source_location: None,
             metadata: serde_json::json!({}),
